requirements.txt
--index-url https://pypi.org/simple
Flask
python-dotenv
Flask-Cors
boto3
PyJWT